Care package package type to enum

+++ LBH.AdultSocialCare.Api/V1/UseCase/Common/Concrete/CreateCarePackageUseCase.cs
@@ -42,7 +42,7 @@ namespace LBH.AdultSocialCare.Api.V1.UseCase.Common.Concrete
             // Get and set random client on package
             var randomClient = await _clientsGateway.GetRandomAsync().ConfigureAwait(false);
             carePackageEntity.ServiceUserId = randomClient.Id;
-            carePackageEntity.PackageType = (int) PackageType.ResidentialCare;
+            carePackageEntity.PackageType = PackageType.ResidentialCare;
             carePackageEntity.Status = PackageStatus.New;
 
             carePackageEntity.ResidentialCareSettings = carePackageSettingsEntity;
